Clean up comments
for some reason it gives about 20% throughput boost (?
